240516 1. 레디스 저장 및 삭제 비동기 코드로 수정 2. 컨트롤러 코드 수정

+++ Controllers/LogoutController.cs
@@ -32,10 +32,18 @@ namespace LoginServer.Controllers
                 return logoutResponse.ErrorCode;
             }
 
-            logoutResponse.ErrorCode = ErrorCode.LogoutSuccess;
-            await _mySqlRepository.DeleteToken(logoutRequest.UserID!);
+            var removeResult = await _redis.RemoveAccount(logoutRequest.UserID!);
+
+            if(removeResult == true)
+            {
+                logoutResponse.ErrorCode = ErrorCode.LogoutSuccess;
+                await _mySqlRepository.DeleteToken(logoutRequest.UserID!);
+            }
+            else
+            {
+                logoutResponse.ErrorCode = ErrorCode.Fail;
+            }            
             
-            _redis.RemoveKey(logoutRequest.UserID!);
             return logoutResponse.ErrorCode;           
         }
     }
